Definition: Business strategy can be understood as the course of action or set of decisions which assist the entrepreneurs in achieving specific business objectives Portfolio page added 1. Portfolio design added 2. database created 3. portfolio create work done
It is nothing but a master plan that the management of a company implements to secure a competitive position in the market, carry on its operations, please customers and achieve the desired ends of the business.

In business, it is the long-range sketch of the desired image, direction and destination of the organization. It is a scheme of corporate intent and action, which is carefully planned and flexibly designed with the purpose of

Achieving effectiveness,
Perceiving and utilizing opportunities,
Mobilising resources,
Securing an advantageous position,
Meeting challenges and threats,
Directing efforts and behaviour and
Gaining command over the situation.
A business strategy is a set of competitive moves and actions that a business uses to attract customers, compete successfully, strengthening performance, and achieve organizational goals. It outlines how business should be carried out to reach the desired ends